backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pdf2image
pdfplumber
Pillow
pytesseract
numpy
rapidfuzz
python-dateutil
python-dotenv
langgraph
langsmith
